Saber

+++ Saber/Saber.Infra/CountWriteOperate.cs
@@ -4,7 +4,7 @@ public class CountWriteOperate : WriteOperate
 {
     public virtual StringValueWrite Write { get; set; }
 
-    public override bool ExecuteChar(uint n)
+    public override bool ExecuteChar(long n)
     {
         long index;
         index = this.Write.Index;
