+++ Saber/Saber.Infra/WriteOperate.cs
@@ -2,7 +2,7 @@ namespace Saber.Infra;
 
 public class WriteOperate : Any
 {
-    public virtual bool ExecuteChar(uint n)
+    public virtual bool ExecuteChar(long n)
     {
         return false;
     }
